Separate base64 implementation from Lua module
