+++ src/nvim/base64.h
@@ -1,0 +1,8 @@
+#ifndef NVIM_BASE64_H
+#define NVIM_BASE64_H
+
+#ifdef INCLUDE_GENERATED_DECLARATIONS
+# include "base64.h.generated.h"
+#endif
+
+#endif  // NVIM_BASE64_H
